Acknowledge file close error only if one happened
The previous version works, but is confusing. In the event that no error results from `f.Close()` and no error resulted during `io.Copy`, it'd set `err` to `nil` (but `err` was already `nil`). The suggested update removes the cause for my confusion and only sets `err` with the error from `f.Close()` in the event that an error actually occurred from that statement, and leaves in place the consideration that no error had occurred from `io.Copy` as well.