Duplicate text and count does not print
Hello - I purchased a digital copy of "The Go Programming Language" and this example does not appear to print the text and count to the console. For reference, I am using version 1.12.7 of the go language. If the logic that iterates over the map of counts is moved inside the standard input scanner then I receive the appropriate log statements.